rename examples to test_examples
