+++ tests/e2e/default_openaispec.py
@@ -1,5 +1,5 @@
 from litserve import OpenAISpec
-from litserve.examples.openai_spec_example import TestAPI
+from litserve.test_examples.openai_spec_example import TestAPI
 import litserve as ls
 
 if __name__ == "__main__":
